+++ src/main/java/com/example/WebSocket/domain/comment.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.Comment;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -14,7 +16,7 @@ import org.hibernate.annotations.Comment;
 public class comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 자동 증가 ID
-    private int comment_id;
+    private BigInteger comment_id;
 
     @Column(nullable = false)
     private String content;
